Agent/server model using MQTT, initial commit. The idea is that multiple agents can be deployed in different parts of the network and be controlled from a central server. Having the agents receive job requests from an MQTT broker should be easier than routing HTTP to the agents.
This commit has
- REST API (server.py) refactored a bit so that the requesthandler subclasses
  mainly handle HTTP and actual business logic is elsewhere, such as the job
  dispatch logic
- the main app starts an mqtt client either in server or agent mode
- note that the mqtt client expects to be supported by an external runloop
  (this is the case currently)
- currently submitting jobs to a server causes them be relayed to an agent

TODO:
- provide an agent list on the GUI
- GUI to have option to start jobs either at server, or at agents or maybe both
- figure out how to access the scan results of agents
